+++ test/fixtures/greek_version.rb
@@ -1,0 +1,5 @@
+module Reissue
+  module GreekGem
+    VERSION = "2.32.beta"
+  end
+end